Add deployment script of rolling node and adapt tezos_node module to handle deployment mode

+++ infra/terraform/modules/tezos_node/user_data.tpl
@@ -12,13 +12,17 @@ curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip
 unzip awscliv2.zip
 ./aws/install
 
+if [ ${mode} == "archive" ]; then
 mkfs.ext4 -E discard /dev/nvme0n1
 mount /dev/nvme0n1 /var/lib/docker/volumes
+fi
 
 mkdir -p ~/.ssh && chmod 700 ~/.ssh
 bash -c '
 echo "${lambda_public_key}" >> /home/ec2-user/.ssh/authorized_keys
 '
+
+if [ ${mode} == "archive" ]; then
 
 cd /var/lib/docker/volumes
 aws s3 cp s3://tzlink-blockchain-data-dev/${network}_node_data.tar.gz ${network}_node_data.tar.gz
@@ -26,14 +30,42 @@ tar xvf ${network}_node_data.tar.gz
 mv archive ${network}_node_data
 chown -R 100:65533 ${network}_node_data
 
+elif [ ${mode} == "rolling" ]; then
+
+aws s3 cp s3://tzlink-blockchain-data-dev/${network}_rolling-snapshot.tar.gz ${network}_rolling-snapshot.tar.gz
+tar xvf ${network}_rolling-snapshot.tar.gz
+else
+
+echo "Error: unknown mode declared. Possible choice : [rolling, archive]"
+
+fi
+
+
+
 curl -o /usr/local/bin/${network}.sh https://gitlab.com/tezos/tezos/raw/${computed_network}/scripts/alphanet.sh
 chmod +x /usr/local/bin/${network}.sh
 
 cd /home/ec2-user
 
-${network}.sh node start --rpc-port 8000 --history-mode archive
+if [ ${mode} == "archive" ]; then
 
+${network}.sh node start --rpc-port 8000 --history-mode archive
 rm -rf /var/lib/docker/volumes/${network}_node_data.tar.gz
+
+elif [ ${mode} == "rolling" ]; then
+
+${network}.sh snapshot import /home/ec2-user/snapshot.rolling
+${network}.sh node start --rpc-port 8000 --history-mode experimental-rolling
+
+rm ${network}_rolling-snapshot.tar.gz
+rm rolling-snapshot.tar.gz
+
+else
+
+echo "Error: unknown mode declared. Possible choice : [rolling, archive]"
+
+fi
+
 
 cat > export-tezos-snap.sh << EOF
 #!/bin/bash -e
